+++ Naotism/Source/Enemy.h
@@ -5,7 +5,7 @@ public:
 public:
 	Enemy::Enemy(int type = 1 , float x = 0 , float y = 0 , int mass = 0);
 	Enemy::Enemy(const Enemy* /*&*/ other);
-	virtual void Enemy::update(double delta);
+	virtual void Enemy::update(float delta);
 	virtual void Enemy::collide( Collidable*& other);
 	Enemy* Enemy::clone()const;
 protected:
